Fix comboData
Now it is requested in DashboardSoli instead of in MisSolicitudes
